refactor(ui): ConfigurableList の props を汎用/特化で型分離しフィルタ UI を抽出
ConfigurableListProps を ListDisplayProps（汎用13: 表示・ページング・状態）と
ListQueryProps（特化9: フィルタ/ソート/検索/URL同期/取得）の合成に分割し、
責務を型シグネチャへ明示。型合成のため呼び出し側は非破壊（apps/web は無変更・
deprecate エイリアス不要）。

フィルタ UI ウィジェット7種 + type 振り分けを configurable-list-filters.tsx
（FilterControl）へ抽出。オーケストレータ本体は 620→270 行に縮小。

コンポーネント自体の2分割（汎用 <List> + <QueryList>）は採らない: 7画面すべてが
fetch+urlSync+sort のフル機構を使い「素の汎用リスト」の消費者がゼロのため
（誰も使わない汎用コンポーネント新設＝過剰一般化を回避）。DS Phase 0 監査 点5。
