+++ packages/ui/src/components/custom/index.ts
@@ -10,7 +10,9 @@ export type {
 	FilterConfig,
 	ListController,
 	ListDataSource,
+	ListDisplayProps,
 	ListError,
+	ListQueryProps,
 	SortConfig,
 	StandardListParams,
 } from "./configurable-list/types";
